Relocate rerun to details.
