+++ client/src/components/History/Content/ContentItem.vue
@@ -304,10 +304,6 @@ function onView() {
     router.push(itemUrls.value.view!);
 }
 
-function onRerun() {
-    router.push(itemUrls.value.rerun!);
-}
-
 function onShowCollectionInfo() {
     router.push(itemUrls.value.showDetails!);
 }
@@ -440,7 +436,6 @@ function unexpandedClick(event: Event) {
                         @view="onView"
                         @showCollectionInfo="onShowCollectionInfo"
                         @edit="onEdit"
-                        @rerun="onRerun"
                         @undelete="onUndelete"
                         @unhide="emit('unhide')" />
                 </span>
