+++ client/src/components/History/Content/Dataset/DatasetActions.vue
@@ -4,6 +4,7 @@ import { faBug, faChartBar, faInfoCircle, faLink, faRedo, faSitemap } from "@for
 import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome";
 import { BButton } from "bootstrap-vue";
 import { computed } from "vue";
+import { useRouter } from "vue-router/composables";
 
 import { type HDADetailed } from "@/api";
 import { copy as sendToClipboard } from "@/utils/clipboard";
@@ -30,14 +31,22 @@ const props = withDefaults(defineProps<Props>(), {
 
 const emit = defineEmits(["toggleHighlights"]);
 
+const router = useRouter();
+
 const showDownloads = computed(() => {
     return !props.item.purged && ["ok", "failed_metadata", "error"].includes(props.item.state);
 });
 const showError = computed(() => {
     return props.item.state === "error" || props.item.state === "failed_metadata";
 });
+const showRerun = computed(() => {
+    return props.item.accessible && props.item.rerunnable && props.item.creating_job && props.item.state != "upload";
+});
 const reportErrorUrl = computed(() => {
     return prependPath(props.itemUrls.reportError!);
+});
+const rerunUrl = computed(() => {
+    return prependPath(props.itemUrls.rerun!);
 });
 const downloadUrl = computed(() => {
     return prependPath(`api/datasets/${props.item.id}/display?to_ext=${props.item.extension}`);
@@ -58,6 +67,10 @@ function onHighlight() {
 
 function onError() {
     window.location.href = reportErrorUrl.value;
+}
+
+function onRerun() {
+    router.push(`/root?job_id=${props.item.creating_job}`);
 }
 </script>
 
@@ -99,6 +112,18 @@ function onError() {
                     variant="link"
                     @click.stop="onHighlight">
                     <FontAwesomeIcon :icon="faSitemap" />
+                </BButton>
+
+                <BButton
+                    v-if="writable && showRerun"
+                    v-b-tooltip.hover
+                    class="rerun-btn px-1"
+                    title="Run Job Again"
+                    size="sm"
+                    variant="link"
+                    :href="rerunUrl"
+                    @click.prevent.stop="onRerun">
+                    <FontAwesomeIcon :icon="faRedo" />
                 </BButton>
             </div>
         </div>
